Use custom column sorting
Even though PDFminer is pretty good at extracting text in reading order,
it sometimes still fails when there is horizontal whitespace across the
whole page.
My algorithm assumes the vertical order is correct, but makes sure the
boxes in the left column come before the columns in the right column.